git-svn-id: file:///svn/nsd/trunk@629 a26ef69c-88ff-0310-839f-98b793d9c207

+++ zparser.c
@@ -1,5 +1,5 @@
 /*
- * $Id: zparser.c,v 1.32 2003/04/22 11:31:51 alexis Exp $
+ * $Id: zparser.c,v 1.33 2003/06/12 12:31:19 erik Exp $
  *
  * zparser.c -- master zone file parser
  *
@@ -1634,13 +1634,13 @@ zprintrdata (FILE *f, int what, u_int16_t *r)
 	case RDATA_PERIOD:
 	case RDATA_LONG:
 		memcpy(&l, &r[1], sizeof(u_int32_t));
-		fprintf(f, "%lu ", ntohl(l));
+		fprintf(f, "%lu ", (unsigned long) ntohl(l));
 		break;
 	case RDATA_SHORT:
-		fprintf(f, "%u ", ntohs(r[1]));
+		fprintf(f, "%u ", (unsigned) ntohs(r[1]));
 		break;
 	case RDATA_BYTE:
-		fprintf(f, "%u ", *((char *)(&r[1])));
+		fprintf(f, "%u ", (unsigned) *((char *)(&r[1])));
 		break;
 	case RDATA_A:
